legacy_python/requirements.txt
pydantic>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
